order_service/requirements.txt
alembic==1.8.1
asyncpg==0.27.0
fastapi==0.88.0
Jinja2==3.1.2
orjson==3.8.3
pydantic==1.10.2
python-dotenv==0.21.0
SQLAlchemy==1.4.45
uvicorn==0.20.0
greenlet==2.0.1
PyJWT==2.6.0
